\name{cauchyreg-package}
\alias{cauchyreg-package}
\docType{package}
\title{
Cauchy Regression
}

\description{
Cauchy regression modelling and LASSO (Least Absolute Shrinkage and Selection Operator) to perform
variable selection are included in this package. Cross-validation is performed to choose the optimal
value of the lambda parameter. LASSO is based on the IRLS (Iteratively Reweighted Least Squares) algorithm.
}

\details{
\tabular{ll}{
Package: \tab cauchyreg\cr
Type: \tab Package\cr
Version: \tab 1.1 \cr
Date: \tab 2026-09-12 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

%\references{
%}





